services/user-service/requirements.txt
django>=5.2,<6
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary
dj-database-url
gunicorn
django-cors-headers
